spilt views.py for readability
